requirements.txt
# Core dependencies
crewai==0.130.0
crewai-tools==0.46.0
chromadb==0.5.23
embedchain==0.1.128
langchain-community>=0.3.1  # Allow a newer version
langchain-core==0.3.62
pydantic==2.8.0
langsmith==0.3.18

# Utilities
python-dotenv==1.0.1
unidecode==1.3.8

# Document & Data processing
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.5
PyYAML==6.0.1
jinja2==3.1.4
graphviz==0.20.3

# Logging
loguru==0.7.2
tenacity==8.3.0